requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
fastapi==0.135.2
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.0
protobuf==6.33.6
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.33.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wrapt==1.17.3
zipp==3.23.0
